+++ src/kustology/bridge.py
@@ -214,15 +214,18 @@ from Kusto.Language.Editor import FormattingOptions, KustoCodeService
 from Kusto.Language.Parsing import TokenParser
 from Kusto.Language.Symbols import (
     ColumnSymbol,
+    CustomReturnType,
     DatabaseSymbol,
     FunctionSymbol,
     Parameter,
     ScalarTypes,
     TableSymbol,
+    Tabularity,
 )
 
 __all__ = [
     "ColumnSymbol",
+    "CustomReturnType",
     "DatabaseSymbol",
     "FormattingOptions",
     "FunctionSymbol",
@@ -233,5 +236,6 @@ __all__ = [
     "Parameter",
     "ScalarTypes",
     "TableSymbol",
+    "Tabularity",
     "TokenParser",
 ]